fix: Ignore redundant assignments in assign_value
assign_value currently stores a new dictionary every time it's
called to assign a specific value, even if the assignment
does not change the dictionary. That is, it saves a dictionary
every time it's called with a value of length 1 even if that value
was already assigned. Ignore redundant assignments to save memory.