Implement Logic for Naked Twins solution
Add functionality to get a list of tuples of the of naked twins. Get
the units that each tuple belongs to . If the value of a single box in
a single unit contains part of the value of one of the twins and is not
one of the twins, then remove said value from the box.